+++ src/openai/types/chat/completion_list_params.py
@@ -15,19 +15,19 @@ class CompletionListParams(TypedDict, total=False):
     """Identifier for the last chat completion from the previous pagination request."""
 
     limit: int
-    """Number of chat completions to retrieve."""
+    """Number of Chat Completions to retrieve."""
 
     metadata: Optional[Metadata]
-    """A list of metadata keys to filter the chat completions by. Example:
+    """A list of metadata keys to filter the Chat Completions by. Example:
 
     `metadata[key1]=value1&metadata[key2]=value2`
     """
 
     model: str
-    """The model used to generate the chat completions."""
+    """The model used to generate the Chat Completions."""
 
     order: Literal["asc", "desc"]
-    """Sort order for chat completions by timestamp.
+    """Sort order for Chat Completions by timestamp.
 
     Use `asc` for ascending order or `desc` for descending order. Defaults to `asc`.
     """