+++ tests/lib/chat/test_completions.py
@@ -58,6 +58,7 @@ ParsedChatCompletion[NoneType](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[NoneType](
+                annotations=None,
                 audio=None,
                 content="I'm unable to provide real-time weather updates. To get the current weather in San Francisco, I
 recommend checking a reliable weather website or app like the Weather Channel or a local news station.",
@@ -126,6 +127,7 @@ ParsedChatCompletion[Location](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[Location](
+                annotations=None,
                 audio=None,
                 content='{"city":"San Francisco","temperature":65,"units":"f"}',
                 function_call=None,
@@ -195,6 +197,7 @@ ParsedChatCompletion[Location](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[Location](
+                annotations=None,
                 audio=None,
                 content='{"city":"San Francisco","temperature":65,"units":"f"}',
                 function_call=None,
@@ -266,6 +269,7 @@ ParsedChoice[ColorDetection](
     index=0,
     logprobs=None,
     message=ParsedChatCompletionMessage[ColorDetection](
+        annotations=None,
         audio=None,
         content='{"color":"red","hex_color_code":"#FF0000"}',
         function_call=None,
@@ -315,6 +319,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":64,"units":"f"}',
             function_call=None,
@@ -329,6 +334,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=1,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":65,"units":"f"}',
             function_call=None,
@@ -343,6 +349,7 @@ def test_parse_pydantic_model_multiple_choices(
         index=2,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content='{"city":"San Francisco","temperature":63.0,"units":"f"}',
             function_call=None,
@@ -393,6 +400,7 @@ ParsedChatCompletion[CalendarEvent](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[CalendarEvent](
+                annotations=None,
                 audio=None,
                 content='{"name":"Science Fair","date":"Friday","participants":["Alice","Bob"]}',
                 function_call=None,
@@ -454,6 +462,7 @@ ParsedChoice[Query](
     index=0,
     logprobs=None,
     message=ParsedChatCompletionMessage[Query](
+        annotations=None,
         audio=None,
         content=None,
         function_call=None,
@@ -565,6 +574,7 @@ def test_parse_pydantic_model_refusal(client: OpenAI, respx_mock: MockRouter, mo
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[Location](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -614,6 +624,7 @@ def test_parse_pydantic_tool(client: OpenAI, respx_mock: MockRouter, monkeypatch
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -686,6 +697,7 @@ def test_parse_multiple_pydantic_tools(client: OpenAI, respx_mock: MockRouter, m
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -767,6 +779,7 @@ def test_parse_strict_tools(client: OpenAI, respx_mock: MockRouter, monkeypatch:
         index=0,
         logprobs=None,
         message=ParsedChatCompletionMessage[NoneType](
+            annotations=None,
             audio=None,
             content=None,
             function_call=None,
@@ -849,6 +862,7 @@ ParsedChatCompletion[Location](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[Location](
+                annotations=None,
                 audio=None,
                 content='{"city":"San Francisco","temperature":58,"units":"f"}',
                 function_call=None,
@@ -924,6 +938,7 @@ ParsedChatCompletion[Location](
             index=0,
             logprobs=None,
             message=ParsedChatCompletionMessage[Location](
+                annotations=None,
                 audio=None,
                 content='{"city":"San Francisco","temperature":65,"units":"f"}',
                 function_call=None,